Move some conversion code out of invoker for less clutter.
